Created more specific Dao-classes
