+++ src/main/java/fi/metropolia/expensetracker/controller/ExpenseController.java
@@ -2,7 +2,9 @@ package fi.metropolia.expensetracker.controller;
 
 import fi.metropolia.expensetracker.MainApplication;
 import fi.metropolia.expensetracker.module.*;
+import fi.metropolia.expensetracker.module.Dao.BudgetExpenseDao;
 import fi.metropolia.expensetracker.module.Dao.Dao;
+import fi.metropolia.expensetracker.module.Dao.SettingsDao;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
@@ -66,6 +68,8 @@ public class ExpenseController {
     private LocalizationManager lan = LocalizationManager.getInstance();
     private Variables variables;
     private Currency currency;
+    private BudgetExpenseDao budgetExpenseDao = new BudgetExpenseDao();
+    private SettingsDao settingsDao = new SettingsDao();
 
 
     public void initialize() {
@@ -130,7 +134,7 @@ public class ExpenseController {
                     if (option.get() == ButtonType.OK) {
                         expenseHistory.getItems().remove(selected);
                         variables.getActiveBudget().removeExpenseFromBudget(selected);
-                        loginSignupDao.deleteExpense(selected.getId());
+                        budgetExpenseDao.deleteExpense(selected.getId());
                         expenseHistory.getItems().clear();
                         expenseHistory.getItems().addAll(variables.getActiveBudget().getExpenses());
 
@@ -169,10 +173,10 @@ public class ExpenseController {
                 ZoneId defaultZoneId = ZoneId.systemDefault();
                 Date finalDate = Date.from(expenseDate.atStartOfDay(defaultZoneId).toInstant());
                 Dao loginSignupDao = new Dao();
-                loginSignupDao.saveExpense(variables.getActiveBudget().getId(), selectTopic.getSelectionModel().getSelectedItem().toString()
+                budgetExpenseDao.saveExpense(variables.getActiveBudget().getId(), selectTopic.getSelectionModel().getSelectedItem().toString()
                         , Double.parseDouble(addExpense.getText()), finalDate);
                 variables.getActiveBudget().resetExpenses();
-                Expense[] expenses = loginSignupDao.getExpenses(variables.getActiveBudget().getId());
+                Expense[] expenses = budgetExpenseDao.getExpenses(variables.getActiveBudget().getId());
                 for (Expense expense : expenses) {
                     variables.getActiveBudget().addExpenseToBudget(expense);
                 }
@@ -224,8 +228,8 @@ public class ExpenseController {
                     constExpenseName.setText(null);
                 } else {
                     Dao loginSignupDao = new Dao();
-                    loginSignupDao.saveConstantExpense(Variables.getInstance().getLoggedUserId(), constExpenseName.getText(), 0.00);
-                    ConstantExpense newConstantExpense = loginSignupDao.getConstantExpenseByName(constExpenseName.getText(), variables.getLoggedUserId());
+                    budgetExpenseDao.saveConstantExpense(Variables.getInstance().getLoggedUserId(), constExpenseName.getText(), 0.00);
+                    ConstantExpense newConstantExpense = budgetExpenseDao.getConstantExpenseByName(constExpenseName.getText(), variables.getLoggedUserId());
                     variables.addConstantExpense(newConstantExpense);
                     selectCategory.setValue(null);
                     constExpense.setText(null);
@@ -249,8 +253,8 @@ public class ExpenseController {
                 ConstantExpense selectedConstExpense = (ConstantExpense) selectCategory.getSelectionModel().getSelectedItem();
                 variables.removeConstantExpense(selectedConstExpense);
                 Dao loginSignupDao = new Dao();
-                loginSignupDao.changeConstantExpenseValue(selectedConstExpense.getId(), Double.parseDouble(constExpense.getText()));
-                ConstantExpense modifiedConstExp = loginSignupDao.getConstantExpenseByName(selectedConstExpense.getType(), variables.getLoggedUserId());
+                budgetExpenseDao.changeConstantExpenseValue(selectedConstExpense.getId(), Double.parseDouble(constExpense.getText()));
+                ConstantExpense modifiedConstExp = budgetExpenseDao.getConstantExpenseByName(selectedConstExpense.getType(), variables.getLoggedUserId());
                 variables.addConstantExpense(modifiedConstExp);
                 variables.setConstExpenses(selectedConstExpense.getType(), Double.parseDouble(constExpense.getText()));
                 selectCategory.setValue(null);
