+++ src/main/java/fi/metropolia/expensetracker/module/Dao/Dao.java
@@ -49,32 +49,6 @@ public class Dao {
         }
     }
 
-    public Integer getAge(int id) throws SQLException {
-
-            PreparedStatement preparedStatement = conn.prepareStatement("SELECT Age FROM Registration WHERE id=?");
-            preparedStatement.setInt(1, id);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return resultSet.getInt(1);
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-            return null;
-    }
-
-    public boolean setAge(int id, int age) throws SQLException {
-       try {
-           PreparedStatement preparedStatement = conn.prepareStatement("UPDATE Registration SET Age = ? WHERE id = ?");
-           preparedStatement.setInt(1, age);
-           preparedStatement.setInt(2, id);
-           preparedStatement.executeUpdate();
-           return true;
-       } catch (SQLException e) {
-           printSQLException(e);
-       }
-       return false;
-    }
 
     public String getPassword(String username) throws SQLException {
 
@@ -130,26 +104,7 @@ public class Dao {
         return null;
     }
 
-    public String loggedThemeColor(Integer id) {
 
-        try {
-            String sql = "SELECT ThemeColor FROM Registration WHERE id = ?";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-
-            preparedStatement.setInt(1, id);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return resultSet.getString(1);
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        return null;
-    }
 
     public boolean userExists(String username) throws SQLException {
         PreparedStatement stmt = conn.prepareStatement("SELECT COUNT(*) FROM Registration WHERE username = ?");
@@ -186,398 +141,6 @@ public class Dao {
         }
     }
 
-    public String loggedCurrency(Integer id) {
-        try {
-            String sql = "SELECT currency FROM Registration WHERE id=?";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
 
-            preparedStatement.setInt(1, id);
 
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return resultSet.getString(1);
-
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        return null;
-    }
-
-    public Budget getBudget(Integer id) {
-        try {
-
-            String sql = "SELECT * FROM Budgets WHERE BudgetId = ?";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-
-            preparedStatement.setInt(1, id);
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return new Budget(resultSet.getInt(1), resultSet.getDouble(3), resultSet.getString(2));
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        return null;
-    }
-
-    public ConstantExpense getConstantExpenseByName(String name, Integer id) {
-        try {
-
-            String sql = "SELECT * FROM Constantexpenses WHERE Title = ? AND registration_id = ?";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-
-            preparedStatement.setString(1, name);
-            preparedStatement.setInt(2, id);
-
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-
-                return new ConstantExpense(resultSet.getInt(1), resultSet.getString(2), resultSet.getDouble(3));
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        return null;
-    }
-
-    public Expense getExpense(Integer id) {
-        try {
-
-            String sql = "SELECT * FROM Expenses WHERE ExpenseId = ?";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-
-            preparedStatement.setInt(1, id);
-
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-
-                return new Expense(resultSet.getInt(1), resultSet.getDouble(3), resultSet.getString(2),
-                        resultSet.getDate(5));
-            }
-            resultSet.close();
-            preparedStatement.close();
-
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        return null;
-    }
-
-    public boolean deleteExpense(Integer id) {
-        Expense expense = getExpense(id);
-
-        if (expense != null) {
-            try {
-                String sql = "DELETE FROM Expenses WHERE ExpenseId = ?";
-                PreparedStatement preparedStatement = conn.prepareStatement(sql);
-                preparedStatement.setInt(1, id);
-                preparedStatement.executeUpdate();
-                return true;
-            } catch (SQLException e) {
-                printSQLException(e);
-                return false;
-            }
-
-        } else {
-            return false;
-        }
-    }
-
-    public void saveConstantExpense(Integer id, String name, Double money) {
-
-        try {
-
-            String sql = "INSERT INTO Constantexpenses VALUES (NULL, ?, ?, ?)";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-            preparedStatement.setString(1, name);
-            preparedStatement.setDouble(2, money);
-            preparedStatement.setInt(3, id);
-
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-    }
-
-    public void saveBudget(Integer id, String name, Double money) {
-
-        try {
-            String sql = "INSERT INTO Budgets VALUES (NULL, ?, ?, ?)";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-            preparedStatement.setString(1, name);
-            preparedStatement.setDouble(2, money);
-            preparedStatement.setInt(3, id);
-
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-    }
-
-    public void saveExpense(Integer id, String type, Double money, Date date) {
-
-        try {
-            String sql = "INSERT INTO Expenses VALUES (NULL, ?, ?, ?, ?)";
-            PreparedStatement preparedStatement = conn.prepareStatement(sql);
-            preparedStatement.setString(1, type);
-            preparedStatement.setDouble(2, money);
-            preparedStatement.setInt(3, id);
-
-
-            java.sql.Date sqlDate = new java.sql.Date(date.getTime());
-            preparedStatement.setDate(4, sqlDate);
-
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-    }
-
-    public Budget[] getBudgets(Integer id) {
-
-        String sql = "SELECT * FROM Budgets WHERE registration_id = ?";
-        Budget[] result;
-        ArrayList<Budget> budgets = new ArrayList<>();
-
-        try {
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, id);
-            ResultSet rs = ps.executeQuery();
-
-            while (rs.next()) {
-                Budget budget = new Budget(rs.getInt(1), rs.getDouble(3), rs.getString(2));
-                budgets.add(budget);
-            }
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        result = new Budget[budgets.size()];
-        for (int i = 0; i < budgets.size(); i++) {
-            result[i] = budgets.get(i);
-        }
-        return result;
-
-    }
-
-    public ConstantExpense[] getConstantExpenses(Integer id) {
-
-        String sql = "SELECT * FROM Constantexpenses WHERE registration_id = ?";
-        ConstantExpense[] result;
-        ArrayList<ConstantExpense> constantExpenses = new ArrayList<>();
-        try {
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, id);
-            ResultSet rs = ps.executeQuery();
-
-            while (rs.next()) {
-                ConstantExpense constantExpense = new ConstantExpense(rs.getInt(1), rs.getString(2), rs.getDouble(3));
-                constantExpenses.add(constantExpense);
-            }
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        result = new ConstantExpense[constantExpenses.size()];
-        for (int i = 0; i < constantExpenses.size(); i++) {
-            result[i] = constantExpenses.get(i);
-        }
-        return result;
-
-    }
-
-    public Expense[] getExpenses(Integer id) {
-        String sql = "SELECT * FROM Expenses WHERE BudgetId = ?";
-        Expense[] result;
-        ArrayList<Expense> expenses = new ArrayList<>();
-
-        try {
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, id);
-            ResultSet rs = ps.executeQuery();
-
-            while (rs.next()) {
-                Expense expense = new Expense(rs.getInt(1), rs.getDouble(3), rs.getString(2),
-                        rs.getDate(5));
-                expenses.add(expense);
-            }
-
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-
-        result = new Expense[expenses.size()];
-        for (int i = 0; i < expenses.size(); i++) {
-            result[i] = expenses.get(i);
-        }
-        return result;
-    }
-
-    public boolean changeUserCurrency(Integer id, String currency) {
-
-        try {
-            String sql = "UPDATE Registration SET currency = ? WHERE id= ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setString(1, currency);
-            ps.setInt(2, id);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean changeUserThemeColor(Integer id, String color) {
-        try {
-            String sql = "UPDATE Registration SET ThemeColor = ? WHERE id = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setString(1, color);
-            ps.setInt(2, id);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean changeConstantExpenseValue(Integer id, Double money) {
-        try {
-            String sql = "UPDATE Constantexpenses SET Amount = ? WHERE ConstantexpenseId = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setDouble(1, money);
-            ps.setInt(2, id);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean changeExpenseMoney(Integer id, Double money) {
-        try {
-            String sql = "UPDATE Expenses SET Money = ? WHERE ExpenseId = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setDouble(1, money);
-            ps.setInt(2, id);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean changeBudgetMoney(Integer id, Double money) {
-
-        try {
-            String sql = "UPDATE Budgets SET Money = ? WHERE BudgetId = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setDouble(1, money);
-            ps.setInt(2, id);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean ModifyBudget(String ogName, Double money, String name) {
-
-        try {
-            String sql = "UPDATE Budgets SET Money = ?, BudgetName = ? WHERE BudgetName = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setDouble(1, money);
-            ps.setString(2, name);
-            ps.setString(3, ogName);
-            ps.executeUpdate();
-            return true;
-        } catch (SQLException e) {
-            printSQLException(e);
-        }
-        return false;
-    }
-
-    public boolean deleteBudget(Integer id) {
-        Budget budget = getBudget(id);
-
-        if (budget != null) {
-            try {
-                String sql = "DELETE FROM Budgets WHERE BudgetId = ?";
-                PreparedStatement preparedStatement = conn.prepareStatement(sql);
-                preparedStatement.setInt(1, id);
-                preparedStatement.executeUpdate();
-                return true;
-            } catch (SQLException e) {
-                printSQLException(e);
-                return false;
-            }
-        } else {
-            return false;
-        }
-    }
-
-    public void deleteUserData(Integer userId) {
-        String deleteIncomes = "DELETE FROM Incomes WHERE UserID = ?";
-        String deleteConstantExpenses = "DELETE FROM Constantexpenses WHERE registration_id = ?";
-        String deleteBudgets = "DELETE FROM Budgets WHERE registration_id = ?";
-        String deleteExpenses = "DELETE FROM Expenses WHERE BudgetId IN (SELECT BudgetId FROM Budgets WHERE registration_id = ?)";
-        String updateUserInfoSql = "UPDATE Registration SET ThemeColor = '#85bb65', currency = 'EUR' WHERE id = ?";
-
-        try {
-            try (PreparedStatement stmt = conn.prepareStatement(deleteIncomes)) {
-                stmt.setInt(1, userId);
-                stmt.executeUpdate();
-            }
-
-            try (PreparedStatement stmt = conn.prepareStatement(deleteConstantExpenses)) {
-                stmt.setInt(1, userId);
-                stmt.executeUpdate();
-            }
-
-            try (PreparedStatement stmt = conn.prepareStatement(deleteBudgets)) {
-                stmt.setInt(1, userId);
-                stmt.executeUpdate();
-            }
-
-            try (PreparedStatement stmt = conn.prepareStatement(deleteExpenses)) {
-                stmt.setInt(1, userId);
-                stmt.executeUpdate();
-            }
-
-            try (PreparedStatement stmt = conn.prepareStatement(updateUserInfoSql)) {
-                stmt.setInt(1, userId);
-                stmt.executeUpdate();
-            }
-
-        } catch (SQLException ex) {
-            ex.printStackTrace();
-        }
-    }
 }
